requirements.txt
numpy
scikit-learn
scipy
tqdm
torch
torchvision
torchmetrics
prdc
pillow
torch-fidelity
matplotlib
tensorflow
opencv-python

backports.functools-lru-cache
cycler
kiwisolver


pyparsing
python-dateutil
pytz

six
subprocess32

ipykernel